requirements.txt
supabase==1.2.0
prometheus-client==0.20.0
pandas==2.2.0
numpy==1.26.0
python-dotenv==1.0.0
faker==22.6.0
psycopg2-binary==2.9.9
pyarrow==15.0.0 